+++ tests/Traits/GroupsTrait/RemoveUserMethodTest.php
@@ -35,4 +35,20 @@ class RemoveUserMethodTest extends TestCase
         self::assertFalse(User::find(2)->hasGroup($this->testUserGroup));
         self::assertFalse(User::find(3)->hasGroup($this->testUserGroup));
     }
+
+    public function test_if_it_can_remove_users_using_array_as_parameter()
+    {
+        $this->testUserGroup->assignUser(1);
+        $this->testUserGroup->assignUser(2);
+        $this->testUserGroup->assignUser(3);
+        self::assertTrue(User::find(1)->hasGroup($this->testUserGroup));
+        self::assertTrue(User::find(2)->hasGroup($this->testUserGroup));
+        self::assertTrue(User::find(3)->hasGroup($this->testUserGroup));
+
+        $this->testUserGroup->removeUser([1, 'User 2', User::find(3)]);
+
+        self::assertFalse(User::find(1)->hasGroup($this->testUserGroup));
+        self::assertFalse(User::find(2)->hasGroup($this->testUserGroup));
+        self::assertFalse(User::find(3)->hasGroup($this->testUserGroup));
+    }
 }